+++ src/Cake.Core/IO/IDirectory.cs
@@ -1,25 +1,18 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace Cake.Core.IO
 {
     /// <summary>
     /// Represents a directory.
     /// </summary>
-    public interface IDirectory
+    public interface IDirectory : IFileSystemInfo
     {
         /// <summary>
         /// Gets the path to the directory.
         /// </summary>
         /// <value>The path.</value>
-        DirectoryPath Path { get; }
-
-        /// <summary>
-        /// Gets a value indicating whether this <see cref="IDirectory"/> exists.
-        /// </summary>
-        /// <value>
-        ///   <c>true</c> if the directory exists; otherwise, <c>false</c>.
-        /// </value>
-        bool Exists { get; }
+        new DirectoryPath Path { get; }
 
         /// <summary>
         /// Creates the directory.
@@ -41,11 +34,49 @@ namespace Cake.Core.IO
         IEnumerable<IDirectory> GetDirectories(string filter, SearchScope scope);
 
         /// <summary>
+        /// Gets directories matching the specified filter and scope, with option to exclude hidden directories
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="scope">The scope.</param>
+        /// <param name="wherePredicate">Filters returned directories based on predicate</param>
+        /// <returns>Directories matching the filter and scope.</returns>
+        IEnumerable<IDirectory> GetDirectories(string filter, SearchScope scope, Func<IFileSystemInfo, bool> wherePredicate);
+
+        /// <summary>
+        /// Gets directories matching the specified filter and scope, with option to exclude hidden directories
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="scope">The scope.</param>
+        /// <param name="wherePredicate">Filters returned directories based on predicate</param>
+        /// <param name="predicateFiltered">Callback if directory gets filtered by wherePredicate</param>
+        /// <returns>Directories matching the filter and scope.</returns>
+        IEnumerable<IDirectory> GetDirectories(string filter, SearchScope scope, Func<IFileSystemInfo, bool> wherePredicate, Action<IFileSystemInfo> predicateFiltered);
+
+        /// <summary>
         /// Gets files matching the specified filter and scope.
         /// </summary>
         /// <param name="filter">The filter.</param>
         /// <param name="scope">The scope.</param>
         /// <returns>Files matching the specified filter and scope.</returns>
         IEnumerable<IFile> GetFiles(string filter, SearchScope scope);
+
+        /// <summary>
+        /// Gets files matching the specified filter and scope.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="scope">The scope.</param>
+        /// <param name="wherePredicate">Filters returned files based on predicate</param>
+        /// <returns>Files matching the specified filter and scope.</returns>
+        IEnumerable<IFile> GetFiles(string filter, SearchScope scope, Func<IFileSystemInfo, bool> wherePredicate);
+
+        /// <summary>
+        /// Gets files matching the specified filter and scope.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="scope">The scope.</param>
+        /// <param name="wherePredicate">Filters returned files based on predicate</param>
+        /// <param name="predicateFiltered">Callback if file gets filtered by wherePredicate</param>
+        /// <returns>Files matching the specified filter and scope.</returns>
+        IEnumerable<IFile> GetFiles(string filter, SearchScope scope, Func<IFileSystemInfo, bool> wherePredicate, Action<IFileSystemInfo> predicateFiltered);
     }
 }